Handle partial writes and stream closing correctly.
When writing to a socket, os.write() does not guarantee that it will
write all bytes supplied. This change adds a buffer on the stream that
keeps track of any pending data that needs to be written, and adds the
stream to the select loop so we know when the stream is writable again.

It also fixes completion behavior of Pumps. Previously, when a Pump hit
EOF on the input stream, it did not close the output stream. That means
that if you ran `echo "hello" | fig run --rm test cat`, it would hang
because the stdin to `cat` would never close. This change will close the
stream after hitting EOF, and the close is properly deferred until after
all bytes are written.

Fixes:

https://github.com/d11wtq/dockerpty/issues/26
https://github.com/docker/compose/issues/658